% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mistakes.R
\name{error}
\alias{error}
\title{Evaluate argument of error/1
(for internal use)}
\usage{
error(expr)
}
\arguments{
\item{expr}{R call of the form error/1}
}
\description{
Evaluate argument of error/1
(for internal use)
}
