% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mistakes.R
\name{expert}
\alias{expert}
\title{Check if a list includes only expert rules
(for internal use)}
\usage{
expert(expr)
}
\arguments{
\item{expr}{list of R calls of the form expert/3 or buggy/3}
}
\description{
Check if a list includes only expert rules
(for internal use)
}
