% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mistakes.R
\name{omit_right}
\alias{omit_right}
\title{Evaluate argument 2 of omit_right/3
(for internal use)}
\usage{
omit_right(expr)
}
\arguments{
\item{expr}{R call of the form omit_right/3}
}
\description{
Evaluate argument 2 of omit_right/3
(for internal use)
}
