% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mistakes.R
\name{pvalue}
\alias{pvalue}
\title{Pretty print pvalue/1}
\usage{
pvalue(x)
}
\arguments{
\item{x}{number 0 < x < 1}
}
\description{
Pretty print pvalue/1
}
\examples{
pvalue(0)
pvalue(0.002)
pvalue(0.062)
pvalue(0.212)
pvalue(1)
}
