% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mistakes.R
\name{;}
\alias{;}
\title{Evaluate arguments of ;/2
(for internal use)}
\usage{
`;`(expr1, expr2)
}
\arguments{
<<<<<<< HEAD
\item{expr1}{R call}

=======
>>>>>>> f658206cfd658c24c136a409a5c638266684f7d7
\item{expr2}{R calls}
}
\description{
Evaluate arguments of ;/2
(for internal use)
}
