% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mistakes.R
\name{feedback}
\alias{feedback}
\title{Extract feedback from a list
(for internal use)}
\usage{
feedback(module, expr)
}
\arguments{
\item{module}{the Prolog module with the feedback}

\item{expr}{the expression
list of R calls of the form expert/3 or buggy/3}
}
\description{
Extract feedback from a list
(for internal use)
}
